Se modularizó la lógica dentro de mercado pago, para cumplir con la modalidad que posee el back
