+++ src/components/mercadoPago/controller.js
@@ -1,41 +1,15 @@
-require('dotenv').config()
-const axios = require('axios')
-const mercadopago = require('mercadopago')
+const Store = require('./store')
 
-const { ACCESS_TOKEN } = process.env
-
-mercadopago.configure({
-  access_token: ACCESS_TOKEN
-})
-
-const getPaymentLink = async (request) => {
-  let paymentLink
-  const preference = {
-    items: request
-  }
-
-  await mercadopago.preferences.create(preference)
-    .then(json => {
-      const editedResponse = {
-        id: json.body.id,
-        url: json.body.init_point
-      }
-      paymentLink = editedResponse
-    })
-
-  return paymentLink
+const getPaymentLink = async () => {
+  return await Store.getPaymentLink()
 }
 
 const getAllPayments = async () => {
-  const response = await axios.get('https://api.mercadopago.com/v1/payments/search', { headers: { Authorization: `Bearer ${ACCESS_TOKEN}` } })
-
-  return response.data
+  return await Store.getAllPayments()
 }
 
 const getPaymentById = async (id) => {
-  const response = await axios.get(`https://api.mercadopago.com/v1/payments/${id}`, { headers: { Authorization: `Bearer ${ACCESS_TOKEN}` } })
-
-  return response.data
+  return Store.getPaymentById(id)
 }
 
 module.exports = {
